requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
jinja2==3.1.4
pydantic==2.9.2
python-multipart==0.0.9
email-validator==2.2.0
httpx==0.27.2
starlette==0.38.6
